[net processing] Don't initialize TxRelay for non-tx-relay peers.
Delay initializing the TxRelay data structure for a peer until we receive
a version message from that peer. At that point we'll know whether it
will ever relay transactions. We only initialize the m_tx_relay
data structure if:

- this isn't an outbound block-relay-only connection; AND
- fRelay=true OR we're not offering NODE_BLOOM to this peer
  (NODE_BLOOM means that the peer may turn on tx relay later)

Note that we will no longer send `feefilter` messages to peers that send
us a `version` message with fRelay=false.