[validation] disconnect m_viewmempool from coins cache when unnecessary
The hierarchy of CoinsViews here is:
m_view -> m_viewmempool -> chainstate coins cache

To prevent DoS, we want to limit access to coins cache and ensure we
uncache coins if the transaction turns out to be invalid.  When
m_viewmempool is always connected to the coins cache, it's harder reason
about this. We could accidentally access the coins cache if we use
m_viewmempool for something.