[mempool] extend CalculateMemPoolAncestors for packages
When calculating ancestor/descendant counts for transactions in the
package, as a heuristic, count every transaction in the package as an
ancestor and descendant of every other transaction in the package.

This may overestimate, but will not underestimate, the
ancestor/descendant counts. Most use cases of submitting as a package
involve two transactions in a parent-child relationship, so this
shortcut still produces an accurate count.