[validation] include vsize in MempoolAcceptResult
Does not change behavior.

In the future, we may want to provide the user with data about both
their transaction and the one in the mempool (identified by wtxid, not
txid). We will want the mempool code to return a result for the
same-txid-different-wtxid transaction that already exists in the mempool
and separately construct a result for that. Thus, we want to use the
vsize from the MempoolAcceptResult rather than the input transaction.