Replace a mempool tx with identical tx with smaller witness
Currently, if a transaction is submitted with the same txid as a
transaction already in the mempool, it's rejected as a duplicate. Change
that behavior to replace the mempool tx if the new tx has a smaller
witness, enough so to satisfy the bip125 (RBF) replacement rules.

But note that this replacement can occur even if the original transaction
is not marked as replaceable. This is because the replacement tx will
pay the same outputs as the tx being replaced (the txids are the same),
so the recipient(s) have no reason to fear not being paid by the
replacement tx.

This commit leverages the existing RBF logic; the same code paths are
followed as with RBF; the only difference is that the tx need not be
marked as replaceable (if txid is the same but wtxid is not).

The approach taken is:

1. remove the replaced tx and all descendent transactions (as in
a normal RBF)

2. replace the transaction

3. add back the descendents.