[refactor] add TxDownloadManager wrapping TxOrphanage and TxRequestTracker
This module is going to be responsible for managing everything related
to transaction download, including txrequest, orphan transactions and
package relay. It will be responsible for managing usage of the
TxOrphanage and instructing PeerManager:
- what tx or package-related messages to send to which peer
- whether a tx or package-related message is allowed or useful
- what transactions are available to try accepting to mempool

Future commits will consolidate the interface and re-delegate most
orphanage and txrequest interactions from PeerManager to TxDownloadManager.