miner: allow CreateNewBlock without non-coinbase txns
There are a few places (in tests) where we want to manually add
transactions to the block template. We currently do this by constructing
an empty mempool for the BlockAssembler to pull transactions from. This
is hacky and potentially unsafe. We should narrow the interface for
constructing new mempools.

Instead of passing in an empty mempool, just add the option to not add
any transactions to the template. In the future, this param be easily
changed to a set of transactions to ask BlockAssembler to force-include.