[policy] define rules for nVersion=3
A new set of rules is needed for contracting protocols in which
presigned transactions between parties are to be fee-bumped at broadcast
time, and need to be resistant to pinning attacks.

In particular, these transactions need to have inherited signaling and
stricter ancestor/descendant limits to prevent RBF (Rule #3) pinning by
attaching large descendants.