[validation] skip last tx in AcceptPackage "submit individually" loop
Prior to this commit, if the last tx is invalid, we needlessly validate
it twice (once in the loop, and then again at the end).

Note that we gate on quit_early=false because otherwise we forget
to populate the results map before returning. This goes away in the next
commit when we pre-fill the results map.