[test] raise wallet_abandonconflict -minrelaytxfee settings
The intention of the test is to set a high -minrelaytxfee
such that these transactions are rejected in LoadMempool() and
in CWallet::ResumbitWalletTransactions().

However, while the parent transactions are below minrelaytxfee, they
each have descendants high enough feerate to bump them past the
minrelaytxfee (observe that the `assert_greater_than` checks fail
if given the original amount of 0.0001). These transactions will be kept
after the mempool persists packages, which is an improvement, but will
cause the original test to fail.