[test] raise wallet_abandonconflict -minrelaytxfee settings
TLDR This test needs to be tweaked because it currently relies on
suboptimal behavior in order to pass, and that behavior is changed in
the next commit.

The intention of the test is to set a high -minrelaytxfee
such that these transactions are rejected in LoadMempool() and
in CWallet::ResumbitWalletTransactions().

However, while the parent transactions are below minrelaytxfee, they
each have descendants high enough feerate to bump them past the
minrelaytxfee (observe that the `assert_greater_than` checks fail
if given the original amount of 0.0001). These transactions will be kept
after the mempool persists packages, causing the original test to fail.