[txorphange] log, erase, and get by wtxid instead of txid
Wtxids are better because they disambiguate between transactions with
the same txid but different witnesses. The TxPackageTracker will work
with wtxids for package relay-related tasks - for example, upon receipt
of an ancpkginfo containing wtxids of transactions, it will look for
those transactions in the orphanage.

This makes it easier to trace a transaction by wtxid through logs in the
category TXPACKAGES.