[validationinterface] add UpdatedBlockTipSync
Allows clients to respond to a new block immediately after it is
connected. This is important for things like recent_rejects, in which
a transaction's validity can change (rejected vs accepted) when this
event is processed (e.g. it has a timelock condition that has just been
met). This is distinct from something like
m_recent_confirmed_transactions in which the validation outcome is the
same (valid vs already-have).