[p2p] limit orphanage by total tx size instead of count
Use the current limit * MAX_STANDARD_TX_WEIGHT, which is the effective
current maximum total orphan weight. As such, as of this commit, the
upper bound has no changed, even though some tests fail because the
orphanage is not trimming as soon as expected. It is because the
transactions are small.

Note: p2p_orphan_handling and p2p_invalid_tx currently fail because they
test that the limit is 100 tranactions.