[txorphanage] trim work sets in AddChildrenToWorkset
Bound the amount of orphans that can be in any peer's work set by
ensuring they never grow beyond a certain point. Since there is limited
space and work sets can contain orphans that have already been deleted,
once per AddChildrenToWorkset call, if we run out of space in a peer's
work set, iterate through it and delete the ones that don't correspond
to a tx in m_orphans.