[p2p] bump DEFAULT_MAX_ORPHAN_ANNOUNCEMENTS to 3000
For the default number of peers (125), allows each to relay a default
descendant package of transactions (up to 25-1=24 can be missing inputs)
out of order.

Functional tests are removed instead of modified because it takes a
very long time to relay 3000 orphans.

Also deletes the now-unused DEFAULT_MAX_ORPHAN_TRANSACTIONS.