[unit test] package validation with (sub)chunking and RBF
- package_ppfp shows benefit of assessing subpackages
- package_ppfc shows that this doesn't let us do "parent pays for
  child;" we only do this when the individual and ancestor feerates meet
mempool minimum feerate
- package_confirmed_parent shows benefit of treating
  txns-already-known as "in mempool" tx
- package_with_rbf tests existing behavior that shouldn't change. Even
  though package RBF is not enabled, it's important that submitting a
transaction as part of a package does not block it from doing a normal
replacement. Otherwise we may blind ourselves to a replacement simply
because it has a child and we happened to download it as a package.
- package_needs_reorder shows necessity of linearizing using cluster
  information.  If we used the original order, we would reject
everything, even if we submit subpackages.
- package_cluster_limits shows that linearization can help us pick the
  most incentive-compatible tx to accept, and that we can take a subchunk
when it is sufficient to meet the minimum feerate.